ProductsMicroservice/DockerFiles/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
sqlalchemy-cockroachdb==2.0.2
requests==2.32.3